// frontend/src/api/generated/model/task.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */
import type { TaskRoadmapItemId } from './taskRoadmapItemId';
import type { TaskStatus } from './taskStatus';
import type { TaskWorkspacePath } from './taskWorkspacePath';

export interface Task {
  created_at: string;
  description: string;
  id: string;
  roadmap_item_id?: TaskRoadmapItemId;
  status: TaskStatus;
  title: string;
  updated_at: string;
  workspace_path?: TaskWorkspacePath;
}